requirements.txt
google-generativeai>=0.8.0
streamlit>=1.28.0